part4/hbnb/requirements.txt
flask
flask-restx
flask-bcrypt
flask-jwt-extended
sqlalchemy
flask-sqlalchemyflask-cors
